Moved all states into states.h and states.cpp Calculate heading to current goal uncertainty during InMission states
