+++ src/bin/mission_manager/CMakeLists.txt
@@ -4,8 +4,8 @@ protobuf_generate_cpp(PROTO_SRCS PROTO_HDRS ${CMAKE_CURRENT_BINARY_DIR} config.p
 
 add_executable(${APP}
   app.cpp
-  machine.cpp
   MissionManager.cpp
+  states.cpp
   ${PROTO_SRCS} ${PROTO_HDRS})
 
 target_link_libraries(${APP}
