+++ src/bin/mission_manager/MissionManagerStateMachine.h
@@ -32,7 +32,6 @@
 #include "jaiabot/messages/imu.pb.h"
 
 #include "events.h"
-#include "states.h"
 
 using namespace jaiabot::protobuf;
 
@@ -46,6 +45,8 @@ class MissionManager;
 
 namespace statechart
 {
+
+struct PreDeployment;
 
 struct MissionManagerStateMachine
     : boost::statechart::state_machine<MissionManagerStateMachine, PreDeployment>,
@@ -309,6 +310,12 @@ struct MissionManagerStateMachine
     }
     const std::string& data_offload_exclude() { return data_offload_exclude_; }
 
+    void set_current_bot_status(const BotStatus& bot_status)
+    {
+        current_bot_status_ = bot_status;
+    }
+    const BotStatus& current_bot_status() { return current_bot_status_; }
+
   private:
     apps::MissionManager& app_;
     jaiabot::protobuf::MissionState state_{jaiabot::protobuf::PRE_DEPLOYMENT__IDLE};
@@ -346,6 +353,7 @@ struct MissionManagerStateMachine
     std::string data_time_string_{""};
     int32_t hub_id_{0};
     std::string data_offload_exclude_{""};
+    BotStatus current_bot_status_;
 };
 
 } // namespace statechart
